+++ src/schema/index.ts
@@ -1,5 +1,4 @@
 export * from './builder';
 export * from './default-data';
-export * from './recursive-builder-helper';
 export * from './type-generator';
 export * from './validation';